chore: reorganize folder structure for pages
